refactor: standardize daemon lifecycle on SpringServiceDaemonSmartLifecycle
Unify all daemon-boot modules on a single lifecycle adapter instead of
three divergent patterns (DaemonSmartLifecycle, SpringServiceDaemonSmartLifecycle,
inline SmartLifecycle anonymous classes).

- Add AbstractServiceDaemon convenience constructor to
  SpringServiceDaemonSmartLifecycle (auto-extracts daemon name)
- Delete redundant DaemonSmartLifecycle (functionally identical for
  AbstractServiceDaemon subclasses since afterPropertiesSet() delegates
  to init() and destroy() delegates to stop())
- Migrate 6 DaemonSmartLifecycle users: Alarmd, EventTranslator,
  Discovery, Pollerd, Enlinkd, Collectd
- Replace 3 inline SmartLifecycle blocks: Provisiond, PerspectivePollerd,
  Telemetryd
- Remove unused LOG fields and Logger imports from PerspectivePollerd
  and Telemetryd configs
- Merge test coverage into single test with @Nested classes

BSMd already used SpringServiceDaemonSmartLifecycle (unchanged).
Syslogd and Trapd are sink consumers with no daemon lifecycle (unchanged).
