+++ core/daemon-common/src/test/java/org/opennms/core/daemon/common/SpringServiceDaemonSmartLifecycleTest.java
@@ -23,54 +23,128 @@ package org.opennms.core.daemon.common;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.opennms.netmgt.daemon.AbstractServiceDaemon;
 import org.opennms.netmgt.daemon.SpringServiceDaemon;
 
 class SpringServiceDaemonSmartLifecycleTest {
 
-    private boolean initCalled = false;
-    private boolean startCalled = false;
-    private boolean destroyCalled = false;
+    @Nested
+    class WithSpringServiceDaemon {
 
-    private final SpringServiceDaemon mockDaemon = new SpringServiceDaemon() {
-        @Override
-        public void afterPropertiesSet() {
-            initCalled = true;
+        private boolean initCalled = false;
+        private boolean startCalled = false;
+        private boolean destroyCalled = false;
+
+        private final SpringServiceDaemon mockDaemon = new SpringServiceDaemon() {
+            @Override
+            public void afterPropertiesSet() {
+                initCalled = true;
+            }
+
+            @Override
+            public void start() {
+                startCalled = true;
+            }
+
+            @Override
+            public void destroy() {
+                destroyCalled = true;
+            }
+        };
+
+        @Test
+        void startCallsAfterPropertiesSetThenStart() {
+            var lifecycle = new SpringServiceDaemonSmartLifecycle(mockDaemon, "TestDaemon");
+            assertThat(lifecycle.isRunning()).isFalse();
+            lifecycle.start();
+            assertThat(initCalled).isTrue();
+            assertThat(startCalled).isTrue();
+            assertThat(lifecycle.isRunning()).isTrue();
         }
 
-        @Override
-        public void start() {
-            startCalled = true;
+        @Test
+        void stopCallsDestroy() {
+            var lifecycle = new SpringServiceDaemonSmartLifecycle(mockDaemon, "TestDaemon");
+            lifecycle.start();
+            lifecycle.stop();
+            assertThat(destroyCalled).isTrue();
+            assertThat(lifecycle.isRunning()).isFalse();
         }
 
-        @Override
-        public void destroy() {
-            destroyCalled = true;
+        @Test
+        void phaseIsMaxValue() {
+            var lifecycle = new SpringServiceDaemonSmartLifecycle(mockDaemon, "TestDaemon");
+            assertThat(lifecycle.getPhase()).isEqualTo(Integer.MAX_VALUE);
         }
-    };
-
-    @Test
-    void startCallsAfterPropertiesSetThenStart() throws Exception {
-        var lifecycle = new SpringServiceDaemonSmartLifecycle(mockDaemon, "TestDaemon");
-        assertThat(lifecycle.isRunning()).isFalse();
-        lifecycle.start();
-        assertThat(initCalled).isTrue();
-        assertThat(startCalled).isTrue();
-        assertThat(lifecycle.isRunning()).isTrue();
     }
 
-    @Test
-    void stopCallsDestroy() throws Exception {
-        var lifecycle = new SpringServiceDaemonSmartLifecycle(mockDaemon, "TestDaemon");
-        lifecycle.start();
-        lifecycle.stop();
-        assertThat(destroyCalled).isTrue();
-        assertThat(lifecycle.isRunning()).isFalse();
-    }
+    @Nested
+    class WithAbstractServiceDaemon {
 
-    @Test
-    void phaseIsMaxValue() {
-        var lifecycle = new SpringServiceDaemonSmartLifecycle(mockDaemon, "TestDaemon");
-        assertThat(lifecycle.getPhase()).isEqualTo(Integer.MAX_VALUE);
+        private static class TestDaemon extends AbstractServiceDaemon {
+            boolean initialized = false;
+            boolean started = false;
+            boolean stopped = false;
+
+            TestDaemon() {
+                super("test-daemon");
+            }
+
+            @Override
+            protected void onInit() {
+                initialized = true;
+            }
+
+            @Override
+            protected void onStart() {
+                started = true;
+            }
+
+            @Override
+            protected void onStop() {
+                stopped = true;
+            }
+        }
+
+        @Test
+        void convenienceConstructorExtractsName() {
+            var daemon = new TestDaemon();
+            var lifecycle = new SpringServiceDaemonSmartLifecycle(daemon);
+            assertThat(lifecycle.isAutoStartup()).isTrue();
+        }
+
+        @Test
+        void startCallsInitThenStart() {
+            var daemon = new TestDaemon();
+            var lifecycle = new SpringServiceDaemonSmartLifecycle(daemon);
+
+            lifecycle.start();
+
+            assertThat(daemon.initialized).isTrue();
+            assertThat(daemon.started).isTrue();
+            assertThat(lifecycle.isRunning()).isTrue();
+        }
+
+        @Test
+        void stopCallsStop() {
+            var daemon = new TestDaemon();
+            var lifecycle = new SpringServiceDaemonSmartLifecycle(daemon);
+
+            lifecycle.start();
+            lifecycle.stop();
+
+            assertThat(daemon.stopped).isTrue();
+            assertThat(lifecycle.isRunning()).isFalse();
+        }
+
+        @Test
+        void isAutoStartupReturnsTrue() {
+            var daemon = new TestDaemon();
+            var lifecycle = new SpringServiceDaemonSmartLifecycle(daemon);
+
+            assertThat(lifecycle.isAutoStartup()).isTrue();
+        }
     }
 }
